add replace text feature
Summary:

Test Plan:
